config employee

+++ src/main/java/hunre/it/demo/repository/EmployeeRepository.java
@@ -1,9 +1,8 @@
 package hunre.it.demo.repository;
 
+import hunre.it.demo.model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import hunre.it.demo.model.Employee;
-
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
 
 }
